requirements.txt
# Core ML & Data Science
scikit-learn>=1.3.0
pandas>=2.0.0
numpy>=1.24.0
scipy>=1.11.0

# Gradient Boosting
xgboost>=1.7.0
catboost>=1.2.0

# Feature Engineering
category-encoders>=2.6.0

# Explainability
shap>=0.42.0

# MLflow
mlflow>=2.5.0

# FastAPI & Web
fastapi>=0.100.0
uvicorn[standard]>=0.22.0
pydantic>=2.0.0

# Authentication & Security
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.6

# Database
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0
alembic>=1.11.0

# Data Validation
great-expectations>=0.17.0

# Monitoring
evidently>=0.2.0

# Utilities
python-dotenv>=1.0.0
requests>=2.31.0
joblib>=1.3.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0

# Development
black>=23.0.0
flake8>=6.0.0
mypy>=1.5.0
